cleaned up services section
